1.修改菜单配置 2.增加禁用名称UI 3.修改链接为Javascript:;状态。

+++ Mozlite.Mvc/Verifiers/Verifiers.cs
@@ -75,7 +75,7 @@ namespace Mozlite.Mvc.Verifiers
 
 
             //定义图像的大小，生成图像的实例  
-            img = new Bitmap((int)code.Length * 18 + 20, 32);
+            img = new Bitmap(code.Length * 18 + 20, 32);
 
             g = Graphics.FromImage(img);//从Img对象生成新的Graphics对象
